diag: Move PrintIfSpace to separate file
